requirements.txt
asgiref==3.2.7
crispy-forms-materialize==0.2
Django==3.0.6
django-crispy-forms==1.9.0
django-materializecss-form==1.1.17
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
mysqlclient==1.3.7